Add Plant overview with status
I want people to be instantly aware if they need to water one of their
plants. We can do this by showing them some kind of progress indicator
that indicates when the plant is dry. In this commit I've added a
circle indicator which a blue color as progress fill color (hinting for
water). As the plant drinks more, the moisture voltage will increase
and the filled area in the indicator will get smaller. So, when all the
blue is gone from the indicator, the plant will have extremely dry dirt!