Improve scrolling behavior for top bar
We want people to be able to return to  have the top bar re-appear when
they scroll up in a list. This adds a piece of JavaScript that makes
the fixed top bar re-appear as soon as a scroll up is detected. It will
hide it as soon as a scroll down is detected.